fix: moved up AWS client folder
