+++ awsClients/ddbDocClient.ts
