+++ utils/entityType.ts
@@ -1,5 +1,5 @@
 import { QueryCommand, QueryCommandInput } from "@aws-sdk/lib-dynamodb";
-import { Dynamo } from "../lib/awsClients/ddbDocClient";
+import { Dynamo } from "../awsClients/ddbDocClient";
 
 const { DYNAMO_TABLE_NAME } = process.env;
