+++ utils/invites/createOrgInvite.ts
@@ -2,7 +2,7 @@ import {
   TransactWriteCommand,
   TransactWriteCommandInput,
 } from "@aws-sdk/lib-dynamodb";
-import { Dynamo } from "../../lib/awsClients/ddbDocClient";
+import { Dynamo } from "../../awsClients/ddbDocClient";
 import { GetAllUserInvites } from "./getAllOrgInvites";
 import { getCurrentTime } from "../time";
 import { nanoid } from "nanoid";
