+++ utils/loginLinks/deleteLoginLink.ts
@@ -1,6 +1,6 @@
 // Deletes the login link the user used to log in
 import { DeleteCommand, DeleteCommandInput } from "@aws-sdk/lib-dynamodb";
-import { Dynamo } from "../../lib/awsClients/ddbDocClient";
+import { Dynamo } from "../../awsClients/ddbDocClient";
 
 const { DYNAMO_TABLE_NAME } = process.env;
 export default async function DeleteLoginLink(
