+++ utils/openings/deleteOpening.ts
@@ -4,7 +4,7 @@ import {
   TransactWriteCommand,
   TransactWriteCommandInput,
 } from "@aws-sdk/lib-dynamodb";
-import { Dynamo } from "../../lib/awsClients/ddbDocClient";
+import { Dynamo } from "../../awsClients/ddbDocClient";
 const { DYNAMO_TABLE_NAME } = process.env;
 import { DeleteStage } from "../stages/deleteStage";
 import { GetAllStagesInOpening } from "../stages/getAllStagesInOpening";
