+++ utils/stages/createStageRule.ts
@@ -1,5 +1,5 @@
 import { PutCommand, PutCommandInput } from "@aws-sdk/lib-dynamodb";
-import { Dynamo } from "../../lib/awsClients/ddbDocClient";
+import { Dynamo } from "../../awsClients/ddbDocClient";
 import { getCurrentTime } from "../time";
 import { nanoid } from "nanoid";
