+++ utils/users/createLoginEvent.ts
@@ -1,7 +1,7 @@
 import { PutCommand, PutCommandInput } from "@aws-sdk/lib-dynamodb";
 import { getCurrentTime, getPastOrFutureTime } from "../time";
 import { getUserByEmail } from "../users/getUserByEmail";
-import { Dynamo } from "../../lib/awsClients/ddbDocClient";
+import { Dynamo } from "../../awsClients/ddbDocClient";
 import { createUser } from "../users/createUser";
 
 const { DYNAMO_TABLE_NAME } = process.env;
